add new dynamo stream
