+++ .github/workflows/main.yml
@@ -14,12 +14,15 @@ jobs:
       - uses: actions/setup-go@v5.5.0
         with:
           go-version: 1.24.1
-      - run: make build-lambda
+      - run: make build-apigateway-lambda
+      - run: make build-dynamostream-lambda
       - run: go version
       - uses: actions/upload-artifact@v4.6.2
         with:
           name: build
-          path: ./handler.zip
+          path: |
+            ./handler.zip
+            ./stream.zip
   deploy_dev:
     runs-on: ubuntu-latest
     environment: dev
